Splitting up downloading and archive handling
