+++ src/main/scala/edu/ie3/simbench/exception/io/ZipperException.scala
@@ -1,0 +1,8 @@
+package edu.ie3.simbench.exception.io
+
+import edu.ie3.simbench.exception.SimbenchException
+
+final case class ZipperException(
+    private val msg: String,
+    private val cause: Throwable = None.orNull
+) extends SimbenchException(msg, cause)
